requirements.txt
streamlit
pandas
matplotlib
wordcloud
numpy
plotly
pillow
python-multipart
azure-storage-blob
altair